Ensure enough entropy available when reading from hardware random generator
Previously, the wait for entropy was implemented in the hardware random number generator
initialisation code. However, the entropy should be checked with each call to fetch a
random word, since data can be read from the hardware at a faster rate than it is
generated.

This matches the linux driver implementation at
https://github.com/torvalds/linux/blob/d4f6d923238dbdb69b8525e043fedef2670d4f2c/drivers/char/hw_random/bcm2835-rng.c